google-ai-agents-intensive/projects/capstone-secure-agent/requirements.txt
# Core AI Framework
google-adk>=0.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data & Analysis
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
requests>=2.31.0

# Code Quality
pylint>=3.0.0
black>=23.10.0
mypy>=1.6.0

# Performance
psutil>=5.9.0

